rename organize.ts file to solidity.ts
